requirements.txt
# Core ML/AI Libraries
mlflow==2.8.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3

# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Configuration & Data Processing
PyYAML==6.0.1

# MLflow Dependencies
cloudpickle==3.0.0
psutil==5.9.6
sqlalchemy==2.0.23

# Optional but recommended for MLflow
boto3==1.34.0
azure-storage-blob==12.19.0
google-cloud-storage==2.10.0

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0

# Jupyter for development (optional)
jupyter==1.0.0
ipykernel==6.27.1

# Environment management
python-dotenv==1.0.0

# HTTP requests (for API testing)
requests==2.31.0

# Data visualization (optional)
matplotlib==3.8.2
seaborn==0.13.0

# Model serialization
joblib==1.3.2

# Security
cryptography==41.0.8

